Fix: Fake battery errors out when sending CAN messages
The fake battery can be used to test the software, without
any battery connected via CAN. In this case, sending any message
over CAN will cause an error because no messages are ACKed
since there is nothing connected. This means the Battery-Emulator
is unable to be used without a battery for testing.

Fix this by not sending any CAN messages when the fake
battery is enabled.

Signed-off-by: Christopher Obbard <obbardc@gmail.com>