Fix: PWM_CONTACTOR_CONTROL now actually engages contactors
When PWM_CONTACTOR_CONTROL was set the contactor coil pins
for positive & negative never actually went high. This was
due to a wrong argument passed to ledcWrite: the existing
code used the PWM channel; while ledcWrite actually expects
the PWM pin to be passed.

Fix this by rewriting all of the calls to ledcWrite to use
the PWM pin rather than the channel. While we are here, fix
some quality-of-life things like ordering of the calls,
using variables instead of hardcoding values and also
guarding PWM_CONTACTOR_CONTROL/GPIO usage in #ifdefs.

Fixes: #472
Signed-off-by: Christopher Obbard <obbardc@gmail.com>